fix. user.service에서 deleteImage와 관련된 로직 보완. fix. challenge repo에서 페이지네이션 관련 함수의 createdAt 오류 수정.

+++ src/users/services/users.service.ts
@@ -105,7 +105,7 @@ export class UserService {
       ? (await this.awsService.uploadImage('outbody_user', file)).key
       : user.imgUrl;
 
-    if (imageUrl) {
+    if (imageUrl && user.imgUrl !== null) {
       await this.awsService.deleteImage(user.imgUrl);
     }
